upload of the script template
You will need to edit the path and example domain for you server config.
This is most likely useless for anyone other than people on a shared hosting account.

It will write to files named http_request.txt and https_requests.txt. For domain names like. example.com and example.com-ssl_log 
from the access-log@ path within you home directory. 

This is just a simple way i found to use a crontab and keep the log file. Some reason it is cleared every 24hrs at 0600 server time and I am interested the the urls that are being visited, this is going to be part of a few project tools I want to try and make. 
For making use out of my shared hosting account and using ssh to build my site.